Refactor the server setup
