--- packages/message-forwarder/src/run.ts
@@ -1,3 +1,0 @@
-import { server } from "./server"
-
-server()
