+++ packages/message-forwarder/src/server.e2e.conductor.test.ts
@@ -15,7 +15,7 @@ import { putIncomingMessageToS3 } from "@moj-bichard7/common/test/s3/putIncoming
 import { randomUUID } from "crypto"
 import fs from "fs"
 import createStompClient from "./createStompClient"
-import { server } from "./server"
+import { messageForwarder } from "./messageForwarder"
 import successExceptionsAHOFixture from "./test/fixtures/success-exceptions-aho.json"
 import successExceptionsPNCMock from "./test/fixtures/success-exceptions-aho.pnc.json"
 
@@ -25,17 +25,12 @@ const mqConfig = createMqConfig()
 const resubmittedAho = fs.readFileSync("src/test/fixtures/success-exceptions-aho-resubmitted.xml").toString()
 const conductorConfig = createConductorConfig()
 
-server()
-
 describe("Server in conductor mode", () => {
   let messageData: string
   let correlationId: string
 
   beforeAll(async () => {
-    await new Promise((resolve) => {
-      client.onConnect = resolve
-      client.activate()
-    })
+    await messageForwarder(client)
   })
 
   afterAll(async () => {
